feature: support getting the location from geoclue
Use dbus to connect to geoclue and get the (approximate) location from
the system. Girouette tries to connect if there is no location set in
the config file or if the location is set to `auto` (in the config or
with `-l/--location`.

This is available under the `geoclue` feature, which is enabled by
default.

It is not really possible to build dbus-rs statically because it depends
on a lot of stuff (systemd among other things), so `geoclue` is not
available when building statically.