feature: add localization support to forecast segments
The language option now requires a language code if the form aa_AA, like
en_US or ja_JP. It defaults to the `LANG` environment variable
otherwise.

The locale is used to print weekday names, and the lang code part is
given to OpenWeather to get location names and weather description in
that language if possible.