fix: do not fail on missing daily/hourly/... forecast data
It was always the intent to not fail to deserialize the api results if
any of the forecast fields are missing, since the availability of data
can vary with location, but somehow the api can either return an empty
array (which was supported) or a missing key (which wasn't).

This makes a missing key the same as an empty forecast array.